+++ social-db/controllers/followprofile.controller.js
@@ -1,42 +1,58 @@
 const model = require("../models");
 
+const Profile = model.profile_model;
 const Followprofile = model.followprofile_model;
 
-function createFollowprofile(req, res) {
-    Followprofile.create({
-        following_Id:req.body.following_Id,
-        profileProfileId: req.body.profileProfileId,
-        followerId: req.body.followerId,
-    }).then((result) => res.json(result));
+Profile.hasMany(Followprofile, {as: "FollowerProfile", foreignKey: "profileId" });
+Followprofile.belongsTo(Profile, {as: "FollowerProfile", foreignKey: "followerId"})
+
+const createFollowprofile = async (req, res) => {
+  try {
+    const followprofile = await Followprofile.create(req.body);
+    return res.status(200).json(followprofile);
+  } catch (error) {
+    return res.status(500).send(error.message);
   }
-  
-  function getById(req, res) {
-    Followprofile.findByPk(req.params.id).then((result) => res.json(result));
-  }
-  
-  async function updateFollowprofile(req, res) {
-    await Followprofile.update(
-      {
-        following_Id:req.body.following_Id,
-        profileProfileId: req.body.profileProfileId,
-        followerId: req.body.followerId,
-      },
-      {
-        where: {
-            following_Id: req.params.id,
-        },
-      }
-    );
-    Followprofile.findByPk(req.params.id).then((result) => res.json(result));
-  }
-  
-  async function deleteFollowprofile(req, res) {
-    await Followprofile.destroy({
+}
+
+const getAllFollowers = async (req, res) => {
+  try {
+    const followprofiles = await Followprofile.findAll({
       where: {
-        following_Id: req.params.id,
+        profileId: req.params.id
       },
+      include: ["FollowerProfile"]
     });
-    Followprofile.findByPk(req.params.id).then((result) => res.json(result));
+    return res.status(200).json({ followprofiles });
+  } catch (error) {
+    return res.status(500).send(error.message);
   }
-  
-  module.exports = { createFollowprofile, getById, updateFollowprofile, deleteFollowprofile };
+}
+
+const getById = async (req, res) => {
+  try {
+    const followprofile = await Followprofile.findByPk(req.params.id, {include: ["FollowerProfile"]});
+    if(followprofile) return res.status(200).json(followprofile);
+
+    throw new Error('Follow Profile not found');
+  } catch (error) {
+    return res.status(500).send(error.message);
+  }
+}
+
+const deleteFollowprofile = async (req, res) => {
+  try {
+    const followprofile = await Followprofile.findByPk(req.params.id);
+    if(followprofile){
+      await Followprofile.destroy({
+        where: { followId: req.params.id }
+      });
+      return res.status(200).send("Follow Profile deleted");
+    }
+    throw new Error("Follow Profile not found");
+  } catch (error) {
+    return res.status(500).send(error.message);
+  }
+};
+ 
+module.exports = { createFollowprofile, getAllFollowers, getById, deleteFollowprofile };
